Wrapping up enemies for now
Started to write code for angel movement but am having a hard time getting it to behave the way I want. I'd like for it to stay elevated but move towards the player to attack while they're in range, but for now I'd like to start constructing levels and better optimizing scripts.